Divided Cmake into 2 binaries: Manual and Autonomous
