+++ apps/rpi5-firmware/CMakeLists.txt
@@ -6,7 +6,7 @@ set(BUILD_TESTS ON)
 set(RPI_BIN_DIR ${CMAKE_BINARY_DIR}/bin)
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 set(SOURCES
-	main.cpp
+	mainRemote.cpp
 )
 
 set(SOURCES_LKA
@@ -31,14 +31,14 @@ add_subdirectory(purePursuit)
 
 if(CMAKE_CROSSCOMPILING)
 	add_subdirectory(comms)
-	#add_executable(${CMAKE_PROJECT_NAME} ${SOURCES})
-	#target_link_libraries(${CMAKE_PROJECT_NAME} PUBLIC comms mode middleware kuksa actuator purePursuit
-	#	opencv_core opencv_imgproc opencv_imgcodecs opencv_highgui opencv_videoio)
+	# Manual Car Executable
+	add_executable(${CMAKE_PROJECT_NAME} ${SOURCES})
+	target_link_libraries(${CMAKE_PROJECT_NAME} PUBLIC comms mode middleware kuksa actuator purePursuit)
+	# Autonomous Car Executable
 	add_executable(Lka ${SOURCES_LKA})
 	target_link_libraries(Lka PUBLIC comms mode middleware kuksa actuator purePursuit
 		opencv_core opencv_imgproc opencv_imgcodecs opencv_highgui opencv_videoio)
-	#install(TARGETS CarControl DESTINATION ${RPI_BIN_DIR})
-	install(TARGETS Lka DESTINATION ${RPI_BIN_DIR})
+	install(TARGETS CarControl Lka DESTINATION ${RPI_BIN_DIR})
 else()
 	if(BUILD_TESTS)
 		add_subdirectory(tests)
